cleanup and final tweaks
