+++ gulp/tasks/watch.js
@@ -3,7 +3,7 @@ var path = require('path');
 var handleErrors = require('../util/handleErrors');
 var paths = require('../../config').paths;
 
-gulp.task('watch', function() { //, ['browser-sync']
+gulp.task('watch', function() {
   // Watch task for images
   gulp.watch(path.join(paths.src + paths.img, '**/*'), ['images'])
     .on('error', handleErrors);
@@ -17,12 +17,12 @@ gulp.task('watch', function() { //, ['browser-sync']
     .on('error', handleErrors);
 
   // Watch tasks for html
-  // gulp.watch(path.join(paths.src + paths.pages, '**/*.html'), ['html-templates'])
-  //   .on('error', handleErrors);
-  gulp.watch(path.join(paths.src + paths.content, '**/*.md'), ['browser-sync-reload']) // ['markdown2'])
+  gulp.watch(path.join(paths.src + paths.pages, '**/*.html'), ['browser-sync-reload'])
     .on('error', handleErrors);
-  gulp.watch(path.join(paths.src + paths.partials, '**/*.html'), ['browser-sync-reload']) // , 'concat'
+  gulp.watch(path.join(paths.src + paths.content, '**/*.md'), ['browser-sync-reload'])
     .on('error', handleErrors);
-  gulp.watch(path.join(paths.src + paths.templates, '**/*.html'), ['browser-sync-reload']) // , 'concat'
+  gulp.watch(path.join(paths.src + paths.partials, '**/*.html'), ['browser-sync-reload'])
+    .on('error', handleErrors);
+  gulp.watch(path.join(paths.src + paths.templates, '**/*.html'), ['browser-sync-reload'])
     .on('error', handleErrors);
 });
