+++ src/partials/_nav.html
@@ -20,9 +20,10 @@
                 {% endif %}
                 {% if section.path === 'articles' %}
                   <ul class="unstyled nav nav--stacked">
-                    {% for articleId, articleObj in siblingData %}
+                    {% for id, articleId in siblingData | sort %}
+                      {% set article = siblingData[articleId] %}
                       <li>
-                        <a class="soft--left{% if relativePath.indexOf(item.path + '/' + section.path + '/' + articleId) !== -1 %} active{% endif %}" href="/{{ item.path }}/{{ section.path }}/{{ articleId }}.html">{{ articleObj.title }}</a>
+                        <a class="soft--left{% if relativePath.indexOf(item.path + '/' + section.path + '/' + articleId) !== -1 %} active{% endif %}" href="/{{ item.path }}/{{ section.path }}/{{ article.fileName }}.html">{{ article.title }}</a>
                       </li>
                     {% endfor %}
                   </ul>
